feat:1.2E10 Flex组件准备与出生机制完善

+++ src/main/java/com/ksptool/ourcraft/sharedcore/network/packets/RequestJoinServerNVo.java
@@ -6,12 +6,12 @@ package com.ksptool.ourcraft.sharedcore.network.packets;
  * 当accepted==1时，会携带sessionId和初始位置信息
  */
 public record RequestJoinServerNVo(
-        int accepted,
-        String reason,
-        Integer sessionId,
-        Double x,
-        Double y,
-        Double z,
-        Float yaw,
-        Float pitch) {}
+        int accepted,  // 0=拒绝, 1=接受
+        String reason, // 拒绝原因
+        Long sessionId, // 玩家本次在世界中的唯一ID
+        Double x, // 玩家出生点X
+        Double y, // 玩家出生点Y
+        Double z, // 玩家出生点Z
+        Float yaw, // 玩家朝向Yaw
+        Float pitch) {} // 玩家朝向Pitch
